Rename the 'X11' platform to 'freedesktop'
This is the first step towards teasing the X11 windowing driver out of
the platform/ directory and into the drivers/ directory.
