+++ platform/freedesktop/export/export.cpp
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  export.h                                                             */
+/*  export.cpp                                                           */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,4 +28,31 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-void register_x11_exporter();
+#include "export.h"
+
+#include "editor/editor_export.h"
+#include "platform/freedesktop/logo.gen.h"
+#include "scene/resources/texture.h"
+
+void register_freedesktop_exporter() {
+
+	Ref<EditorExportPlatformPC> platform;
+	platform.instance();
+
+	Ref<Image> img = memnew(Image(_freedesktop_logo));
+	Ref<ImageTexture> logo;
+	logo.instance();
+	logo->create_from_image(img);
+	platform->set_logo(logo);
+	platform->set_name("Linux/X11");
+	platform->set_extension("x86");
+	platform->set_extension("x86_64", "binary_format/64_bits");
+	platform->set_release_32("linux_x11_32_release");
+	platform->set_debug_32("linux_x11_32_debug");
+	platform->set_release_64("linux_x11_64_release");
+	platform->set_debug_64("linux_x11_64_debug");
+	platform->set_os_name("X11");
+	platform->set_chmod_flags(0755);
+
+	EditorExport::get_singleton()->add_export_platform(platform);
+}
