+++ platform/freedesktop/logo.png
