+++ platform/freedesktop/power_x11.cpp
